refactory user logic as a single package
