+++ src/main/java/com/springboot/webfluxapi/user/IRepository.java
@@ -1,0 +1,6 @@
+package com.springboot.webfluxapi.user;
+
+import org.springframework.data.mongodb.repository.ReactiveMongoRepository;
+
+public interface IRepository extends ReactiveMongoRepository<User, String> {
+}
