+++ src/main/java/com/springboot/webfluxapi/user/User.java
@@ -1,4 +1,4 @@
-package com.springboot.webfluxapi.document;
+package com.springboot.webfluxapi.user;
 
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
